bhhh

+++ routes/indexRouter.js
@@ -23,11 +23,25 @@ router.get('/event/:day/:session/:lang?', async (req, res, next)=> {
 });
 
 router.get('/admin', function(req, res, next) {
+  if(!req.session.user)
+    res.render('adminLogin', { title: 'admin' });
+  else
   res.render('admin', { title: 'admin' });
+});
+router.post('/admin', function(req, res, next) {
+  if(req.body.login="editor" && req.body.password=="bolero123"){
+    req.session.user=true;
+    res.render('admin', { title: 'admin' });
+  }
+  else
+    res.render('adminLogin', { title: 'admin' });
 });
 
 
 router.post('/day', async (req, res, next)=> {
+
+  if(!req.session.user)
+    return res.sendStatus(403)
   console.log(req.body.key);
   await req.knex("t_confDays").insert({key:req.body.key});
   res.json(await req.knex.select("*").from("t_confDays"));
@@ -36,6 +50,8 @@ router.get('/day', async (req, res, next)=> {
   res.json(await req.knex.select("*").from("t_confDays"));
 });
 router.post('/session', async (req, res, next)=> {
+  if(!req.session.user)
+    return res.sendStatus(403)
   await req.knex("t_confSession").insert({key:req.body.key, dayid:req.body.id});
   res.json(await req.knex.select("*").from("t_confSession").where({dayid:req.body.id}).orderBy("key"));
 });
@@ -43,6 +59,8 @@ router.get('/session/:id', async (req, res, next)=> {
   res.json(await req.knex.select("*").from("t_confSession").where({dayid:req.params.id}).orderBy("key"));
 });
 router.post('/sessionChange', async (req, res, next)=> {
+  if(!req.session.user)
+    return res.sendStatus(403)
   var id=req.body.item.id;
   delete req.body.item.id;
   await req.knex("t_confSession").update(req.body.item).where({id:id});
@@ -50,7 +68,8 @@ router.post('/sessionChange', async (req, res, next)=> {
 });
 
 router.delete('/session/:id', async (req, res, next)=> {
-
+  if(!req.session.user)
+    return res.sendStatus(403)
   await req.knex("t_confSession").where({id:req.params.id}).delete();
   res.json(req.params.id);
 });
